+++ bin/args.js
@@ -1,0 +1,8 @@
+module.exports = {
+  input: {
+    name: 'input',
+    type: 'string',
+    required: true,
+    description: 'css file to use as input'
+  }
+}