+++ bin/commands/bundle.js
@@ -1,35 +1,37 @@
-var fs = require('fs')
-var csskit = require('../../index')()
+const fs = require('fs')
+const csskit = require('../../index')()
 
-module.exports = {
-  name: 'bundle',
-  command: function bundle (args) {
-    args.input = args._[0]
+const { output, minify } = require('../flags')
+const { input } = require('../args')
 
-    var css = fs.readFileSync(args.input)
+function command (args, flags, context) {
+  const { input } = args
+  const { output, minify } = flags
 
-    csskit.bundle(css, args, function (err, result) {
-      if (err) return console.log(err)
+  const css = fs.readFileSync(input)
 
-      if (args.output) {
-        return fs.writeFileSync(args.output, result.css)
-      }
+  csskit.bundle(css, { input, minify }, function (err, result) {
+    if (err) return console.error(err)
 
-      console.log(result.css)
-    })
-  },
-  options: [
+    if (output) {
+      return fs.writeFileSync(output, result.css)
+    }
+
+    console.log(result.css)
+  })
+}
+
+const args = [input]
+const flags = [output, minify]
+
+const options = {
+  description: 'Create a css bundle',
+  examples: [
     {
-      name: 'output',
-      abbr: 'o',
-      default: null,
-      help: 'specify an output file for the bundled css'
-    },
-    {
-      name: 'minify',
-      abbr: 'm',
-      default: null,
-      help: 'minify using npmjs.com/cssnano'
+      cmd: 'csskit bundle style.css > bundle.css',
+      description: 'Create a css bundle from a source css file'
     }
   ]
 }
+
+module.exports = { command, args, flags, options }